+++ Common.Build/Config/SolutionConfig.cs
@@ -2,8 +2,8 @@ using JetBrains.Annotations;
 
 namespace Common.Build.Config;
 
-[UsedImplicitly]
 public class SolutionConfig
 {
     public required string Name { get; [UsedImplicitly] init; }
+    public required string? DefaultProject { get; [UsedImplicitly] init; }
 }